feat: control swipe delete only when swipe all the way
